webscan_v2/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic[email]>=2.6.0
python-dotenv>=1.0.0
